Add JSON validation to post update method
